# openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Medusa Digital Downloads API
  version: 0.3.0
  summary: Native Medusa v2 APIs for protected digital delivery and licensing.
  description: |
    Admin catalog, storage, entitlement, audit, and reporting APIs plus Store APIs
    for previews, customer libraries, opaque guest capabilities, protected content,
    and key-based license activation. Raw storage locations, hashes, credentials,
    and encrypted material are never part of a response schema.
  contact:
    name: MakePay
    url: https://makepay.io
  license:
    name: MIT
servers:
  - url: http://localhost:9000
    description: Local Medusa server
...
paths:
...
  /admin/digital-downloads/entitlements:
    get:
      operationId: listDigitalEntitlements
      tags: [Admin entitlements]
      summary: List digital entitlements
      security:
        - adminBearer: []
        - adminSession: []
        - adminApiKey: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/Order"
        - $ref: "#/components/parameters/Search"
        - name: order_id
          in: query
          schema:
            $ref: "#/components/schemas/Id"
        - name: customer_id
          in: query
          schema:
            $ref: "#/components/schemas/Id"
        - name: product_config_id
          in: query
          schema:
            $ref: "#/components/schemas/Id"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EntitlementStatus"
      responses:
        "200":
          $ref: "#/components/responses/AdminEntitlementPage"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminSession:
      type: apiKey
      in: cookie
      name: connect.sid
    adminApiKey:
      type: apiKey
      in: header
      name: x-medusa-access-token
    customerBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    customerSession:
      type: apiKey
      in: cookie
      name: connect.sid
    publishableKey:
      type: apiKey
      in: header
      name: x-publishable-api-key
    grantBearer:
      type: http
      scheme: bearer
      bearerFormat: Opaque digital download grant
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    Order:
      name: order
      in: query
      description: Sort field; prefix with `-` for descending order.
      schema:
        type: string
        minLength: 1
        maxLength: 64
    Search:
      name: q
      in: query
      schema:
        type: string
        minLength: 1
        maxLength: 200
...
  responses:
    Error:
      description: Safe error response; internal details are redacted.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    AdminEntitlementPage:
      description: Paginated admin entitlement projections with secret material removed.
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/Page"
              - type: object
                required: [entitlements]
                properties:
                  entitlements:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminEntitlement"
...
  schemas:
    Id:
      type: string
      pattern: "^[A-Za-z0-9][A-Za-z0-9_.:-]{0,127}$"
      maxLength: 128
    Page:
      type: object
      required: [count, limit, offset]
      properties:
        count:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
          maximum: 100
        offset:
          type: integer
          minimum: 0
    Error:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
        code:
          type: string
        message:
          type: string
          maxLength: 500
        request_id:
          type: string
        details:
          type: object
          additionalProperties: true
...
    EntitlementStatus:
      type: string
      enum: [pending, active, expired, suspended, revoked, refunded]
...
    AdminEntitlement:
      type: object
      description: Admin-only operational entitlement projection with secret material recursively removed.
      required: [id, status]
      additionalProperties: true
      properties:
        id:
          $ref: "#/components/schemas/Id"
        status:
          $ref: "#/components/schemas/EntitlementStatus"
        order_id:
          type: [string, "null"]
        line_item_id:
          type: [string, "null"]
        product_config_id:
          type: [string, "null"]
        product_id:
          type: [string, "null"]
        variant_id:
          type: [string, "null"]
        customer_id:
          type: [string, "null"]
        title:
          type: [string, "null"]
        quantity_unit:
          type: [integer, "null"]
          minimum: 0
        granted_at:
          type: [string, "null"]
          format: date-time
        expires_at:
          type: [string, "null"]
          format: date-time
        revoked_at:
          type: [string, "null"]
          format: date-time
        revocation_reason:
          type: [string, "null"]
        licenses:
          type: array
          items:
            type: object
            additionalProperties: true
...